refactor: seperate business logic
